requirements.txt
asgiref==3.8.1
Django==5.0.6
django-crontab==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
PyYAML==5.3.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
